Setup Java Demo
The demo uses an embedded HSQLDB in server mode.
The reason to use the server mode was that debugging of an in-memory
instance is quite hard because you can't connect to a running in-memory
instance because a new connection would also create a new database
instsance. Existing demo tables are dropped to ensure that the data is
fresh on each start.
While implementing the demo I faced one big problem and that was
comparing unix timestamps. One might think that this is easy because
everything is UTC but it in practice it is not as easy as that.
The demo table previously had a column of type `TIMESTAMP` of which
values were checked against the given token timestamp by this WHERE
clause: `UNIX_TIMESTAMP(timestamp) >= token_timestamp`
The problem here is not obvious, token_timestamp contained the epoch
seconds of a timestamp in the local time of my machine, i.e. the epoch
seconds until some time X in CET (UTC+1). Now, converting the tables
timestamp column to a `UNIX_TIMESTAMP` caused it to get an 1h (3600s)
offset so it can be compared to the token_timestamp. But, the timestamp
already was in the timezone of my local machine, so there was no need to
add this offset.
The solution was to change the columns type to `TIMESTAMP WITH TIME
ZONE` so it was aware of the zone and must not add any offset
whatsoever.